Parts combination (#15)
* Created images for rf range buttons

* Deleted unnecessary files

* Updated main_window ui

* Added jammer fields to settings

* Moved LogService in params for LogDialog

* Populated test_data directory

* Seted default value for timestamp in LogEntry

* Added functionality to LogService for real-time logging

* Updated MapService for correctly map moving

* Updated test_data

* Updated .gitignore

* Updated main_window ui

* Refactored and added parts functionality to MainWindow

* Added bg to main_window map ui

* Added new jammer field to settings

* Created GPSData model

* Connected GPSData model to network services

* Moved map view logic into specific class

* Moved jammer logic into specific service

* Moved radar renderer logic into specific class

* Moved calculate_distance gps into specific util

* Connected new specific parts to MainWindow

+++ app/services/settings_service.py
@@ -42,7 +42,7 @@ class SettingsService(QObject):
     # Signal
     radio_range_GHz: list[float]
     # Timers
-    gps_interval_s: int
+    gps_interval_s: float
     # Jammer
     main_relays: list[str]
     is_jammer_auto_start_enabled: bool
